requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
numpy
pydantic
pyjwt
python-jose[cryptography] 
passlib
sqlalchemy
load_dotenv
asyncpg
black
python-dotenv
bcrypt